Fix crash with nil XPath variables
Prior to d91e124ab9843d2eaf62071bf029301254c36e2b, an XPath variable
could be set to nil, in which case it would be silently converted to an
empty string. This is clearly still the intended behaviour of this
method, as can be seen from the `when nil` further down. But, with the
introduction of context support, this case was broken, because trying
to do @@context[:node] would raise a NoMethodError if there was no
context.