Fix a problem that parse exception message can't be generated for invalid encoding XML (#123)
## Why?

If the XML tag contains Unicode characters and an error is occurred for
the tag, an incompatible encoding error is raised. Because our parse
exception message parts have an UTF-8 part (that includes the target tag
information) and an ASCII-8BIT part (that includes error context input).

Fix GH-29

Reported by DuKewu. Thanks!!!