Improve `Element#attribute` as 6500x faster
`Element#namespaces` is heavy method because this method needs to
traverse all ancestors of the element. `Element#attribute` calls
`namespaces` redundantly, so it is much slower.

This commit reduces `namespaces` calls in `Element#attribute`.
Also, this commit removes a redundant `respond_to?` because
`namespaces` must return `Hash` in the current implementation.