Handle unprefixed attributes in `Element#attribute` correctly
Currently, `Element#attrubute` handles unprefixed attributes as
attributes with the default namespace, and handles the default
namespace as `nil` value. However, it is a mis-interpretation of
the specification of XML 1.0.

XML 1.0 specification says:
(https://www.w3.org/TR/xml-names/#defaulting)

> The namespace name for an unprefixed attribute name always has
> no value.

Therefore, we need to distinguish `nil` and the default namespace.
Also, unprefixed attributes should not be handled as attributes
with the default namespace.

Note that XML 1.0 specification also says:

> Default namespace declarations do not apply directly to attribute
> names; the interpretation of unprefixed attributes is determined
> by the element on which they appear.

This means that we can know (programmatically) how to handle
unprefixed attributes from the element's namespace. It does not mean that
the namespace of an unprefixed attribute becomes the default namespace.

This commit changes `Element#attribute` behavior correctly and updates
tests for it.