Fix performance issue caused by using repeated `>` characters inside `<?xml`
A `<` is treated as a string delimiter. 
In certain cases, if `<` is used in succession, read and match are repeated, which slows down the process.
Therefore, the following is used to read ahead to a specific part of the string in advance.

## Proof of Concept
```ruby
require "rexml"
require "benchmark"

def test(benchmark, payload)
  benchmark.report { begin REXML::Document.new(payload) rescue Exception end }
end

Benchmark.bm do |x|
  test(x, '<?xml version="1.0" ' + ">" * 20000 + ' ?>')
  test(x, '<?xml version="1.0" ' + ">" * 40000 + '?>')
  test(x, '<?xml version="1.0" ' + ">" * 60000 + '?>')
  test(x, '<?xml version="1.0" ' + ">" * 80000 + '/>')
  test(x, '<?xml version="1.0" ' + ">" * 100000 + '/>')
end
```